💄 Adapt sidebar to look better on mobile
Before this change, the full sidebar was rendered on the top of the page when viewing it on mobile devices. I could not even see the headline of the post I was reading because my screen was filled with the sidebar.

This change shrinks the contents of the sidebar such that the profile picture and the name is rendered next to each other. Additionally, only the post categories are shown next to each other (e.g. About, Posts). Finally, the footer (incl. socials etc) is displayed at the bottom of the page.